Add screener creation capabilities for admins

+++ project/src/main/java/edu/depaul/cdm/se452/project/model/CourseRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 
 //repository class to help perform CRUD operations on the courses table
-public interface CourseRepository extends CrudRepository<Course, Long> {
+public interface CourseRepository extends CrudRepository<Course, Integer> {
 	//find a course by passing a course number
 	List<Course> findByCourseNumber(int num);
